Some changes I forgot from dev-unstable
