+++ controllers/ppo/ppo.py
@@ -1,0 +1,167 @@
+"""Controller using a PPO agent trained with stable-baselines3."""
+
+from __future__ import annotations  # Python 3.10 type hints
+
+import numpy as np
+from stable_baselines3 import PPO
+
+from lsy_drone_racing.command import Command
+from lsy_drone_racing.controller import BaseController
+from lsy_drone_racing.wrapper import obs_transform
+
+
+class Controller(BaseController):
+    """Template controller class."""
+
+    def __init__(
+        self,
+        initial_obs: np.ndarray,
+        initial_info: dict,
+        buffer_size: int = 100,
+        verbose: bool = False,
+    ):
+        """Initialization of the controller.
+
+        INSTRUCTIONS:
+            The controller's constructor has access the initial state `initial_obs` and the a priori
+            infromation contained in dictionary `initial_info`. Use this method to initialize
+            constants, counters, pre-plan trajectories, etc.
+
+        Args:
+            initial_obs: The initial observation of the environment's state. Consists of
+                [drone_xyz_yaw, gates_xyz_yaw, gates_in_range, obstacles_xyz, obstacles_in_range,
+                gate_id]
+            initial_info: The a priori information as a dictionary with keys 'symbolic_model',
+                'nominal_physical_parameters', 'nominal_gates_pos_and_type', etc.
+            buffer_size: Size of the data buffers used in method `learn()`.
+            verbose: Turn on and off additional printouts and plots.
+        """
+        super().__init__(initial_obs, initial_info, buffer_size, verbose)
+        # Save environment and control parameters.
+        self.CTRL_TIMESTEP = initial_info["ctrl_timestep"]
+        self.CTRL_FREQ = initial_info["ctrl_freq"]
+        self.initial_obs = initial_obs
+        self.VERBOSE = verbose
+        self.BUFFER_SIZE = buffer_size
+
+        # Store a priori scenario information.
+        self.NOMINAL_GATES = initial_info["nominal_gates_pos_and_type"]
+        self.NOMINAL_OBSTACLES = initial_info["nominal_obstacles_pos"]
+
+        # Reset counters and buffers.
+        self.reset()
+        self.episode_reset()
+
+        self.model_name = "models/ppo_2024-05-16_23-18-32"
+        self.model = PPO.load(self.model_name)
+
+    def compute_control(
+        self,
+        ep_time: float,
+        obs: np.ndarray,
+        reward: float | None = None,
+        done: bool | None = None,
+        info: dict | None = None,
+    ) -> tuple[Command, list]:
+        """Pick command sent to the quadrotor through a Crazyswarm/Crazyradio-like interface.
+
+        INSTRUCTIONS:
+            Re-implement this method to return the target position, velocity, acceleration,
+            attitude, and attitude rates to be sent from Crazyswarm to the Crazyflie using, e.g., a
+            `cmdFullState` call.
+
+        Args:
+            ep_time: Episode's elapsed time, in seconds.
+            obs: The environment's observation [drone_xyz_yaw, gates_xyz_yaw, gates_in_range,
+                obstacles_xyz, obstacles_in_range, gate_id].
+            reward: The reward signal.
+            done: Wether the episode has terminated.
+            info: Current step information as a dictionary with keys 'constraint_violation',
+                'current_target_gate_pos', etc.
+
+        Returns:
+            The command type and arguments to be sent to the quadrotor. See `Command`.
+        """
+        new_obs = obs_transform(obs=obs, info=info)
+        action, next_predicted_state = self.model.predict(
+            new_obs, deterministic=True)
+
+        # Prepare the command to be sent to the quadrotor.
+        command_type = Command.FULLSTATE
+        x = float(action[0])
+        y = float(action[1])
+        z = float(action[2])
+        target_pos = np.array([x, y, z])
+        target_vel = np.zeros(3)
+        target_acc = np.zeros(3)
+        target_yaw = float(action[3])
+        target_rpy_rates = np.zeros(3)
+        args = [target_pos, target_vel, target_acc,
+                target_yaw, target_rpy_rates, ep_time]
+
+        # logger.info(f"Predicted action: {action}")
+        # logger.info(f"Args: {args}")
+        return command_type, args
+
+    def step_learn(
+        self,
+        action: list,
+        obs: np.ndarray,
+        reward: float | None = None,
+        done: bool | None = None,
+        info: dict | None = None,
+    ):
+        """Learning and controller updates called between control steps.
+
+        INSTRUCTIONS:
+            Use the historically collected information in the five data buffers of actions,
+            observations, rewards, done flags, and information dictionaries to learn, adapt, and/or
+            re-plan.
+
+        Args:
+            action: Most recent applied action.
+            obs: Most recent observation of the quadrotor state.
+            reward: Most recent reward.
+            done: Most recent done flag.
+            info: Most recent information dictionary.
+
+        """
+        #########################
+        # REPLACE THIS (START) ##
+        #########################
+
+        # Store the last step's events.
+        self.action_buffer.append(action)
+        self.obs_buffer.append(obs)
+        self.reward_buffer.append(reward)
+        self.done_buffer.append(done)
+        self.info_buffer.append(info)
+
+        # Implement some learning algorithm here if needed
+
+        #########################
+        # REPLACE THIS (END) ####
+        #########################
+
+    def episode_learn(self):
+        """Learning and controller updates called between episodes.
+
+        INSTRUCTIONS:
+            Use the historically collected information in the five data buffers of actions,
+            observations, rewards, done flags, and information dictionaries to learn, adapt, and/or
+            re-plan.
+
+        """
+        #########################
+        # REPLACE THIS (START) ##
+        #########################
+
+        _ = self.action_buffer
+        _ = self.obs_buffer
+        _ = self.reward_buffer
+        _ = self.done_buffer
+        _ = self.info_buffer
+
+        #########################
+        # REPLACE THIS (END) ####
+        #########################
