+++ pyproject.toml
@@ -50,7 +50,6 @@ exclude = [
     "site-packages",
     "venv",
 ]
-
 line-length = 100
 indent-width = 4
 target-version = "py38"
@@ -71,7 +70,7 @@ suppress-none-returning = true
 [tool.ruff.format]
 quote-style = "double"
 indent-style = "space"
-skip-magic-trailing-comma = true
+skip-magic-trailing-comma = false
 line-ending = "auto"
 docstring-code-format = true
 docstring-code-line-length = "dynamic"
